split up inviter component
